Original

+++ SuiviDesEtudiants/app/Http/Controllers/UserController.php
@@ -6,6 +6,9 @@ use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rules;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\ValidationException;
+
 
 class UserController extends Controller
 {
@@ -36,29 +39,36 @@ class UserController extends Controller
 
     public function update(Request $request, $id)
     {
+        $currentUser = Auth::user();
+
+        if ($currentUser && $currentUser->id == $id) {
+            return redirect()->route('gestion-des-utilisateurs')->withErrors(['error' => 'Vous ne pouvez pas modifier votre propre compte.']);
+        }
         $request->validate([
-            'matricule' => 'required|string|min:5|max:5|unique:users,Matricule,' . $id,
+            'Matricule' => 'required|string|min:5|max:5|unique:users,Matricule,' . $id,
             'role' => 'required',
-            'email' => 'required|email|unique:users,Email,' . $id,
-            'password' => 'nullable|min:8|confirmed',
+            'Email' => 'required|email|unique:users,Email,' . $id,
         ]);
 
         $user = User::findOrFail($id);
-        $user->update([
-            'Matricule' => $request->matricule,
-            'role' => $request->role,
-            'Email' => $request->email,
-            'password' => $request->password ? Hash::make($request->password) : $user->password,
-        ]);
+        $user->Matricule = $request->Matricule;
+        $user->role = $request->role;
+        $user->Email = $request->Email;
+        $user->save();
 
         return redirect()->route('gestion-des-utilisateurs')->with('success', 'Utilisateur mis à jour avec succès.');
     }
 
     public function destroy($id)
     {
+        $currentUser = Auth::user();
+
+        if ($currentUser && $currentUser->id == $id) {
+            return redirect()->route('gestion-des-utilisateurs')->withErrors(['error' => 'Vous ne pouvez pas supprimer votre propre compte.']);
+        }
         $user = User::findOrFail($id);
         $user->delete();
-
         return redirect()->route('gestion-des-utilisateurs')->with('success', 'Utilisateur supprimé avec succès.');
     }
+
 }
